+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [18.15.6] - 2021-XX-XX
 ### Changed
 - removed shipped libresample archive used for building, and using now the official centos repository package
+- fixed /etc/aliases management
 
 ## [18.15.5] - 2021-03-18
 ### Changed